+++ test/server/proxy/https-test.js
@@ -1,0 +1,75 @@
+const selfSignedCertificate = require('openssl-self-signed-certificate');
+const { expect }            = require('chai');
+const request               = require('request-promise-native');
+const fs                    = require('fs');
+
+const {
+    createSession,
+    createProxy,
+    compareCode,
+    createDestinationServer
+} = require('./utils');
+
+const {
+    PAGE_ACCEPT_HEADER,
+    SAME_DOMAIN_SERVER_PORT
+} = require('./constants');
+
+describe('https proxy', () => {
+    let session    = null;
+    let proxy      = null;
+    let destServer = null;
+
+    before(() => {
+        const sameDomainDestinationServer = createDestinationServer(SAME_DOMAIN_SERVER_PORT);
+        const { app }                     = sameDomainDestinationServer;
+
+        destServer = sameDomainDestinationServer.server;
+
+        app.get('/page', (req, res) => {
+            res.setHeader('content-type', 'text/html');
+            res.end(fs.readFileSync('test/server/data/page/src.html').toString());
+        });
+    });
+
+    after(() => {
+        destServer.close();
+    });
+
+    beforeEach(() => {
+        session = createSession();
+
+        proxy = createProxy({
+            key:  selfSignedCertificate.key,
+            cert: selfSignedCertificate.cert
+        });
+    });
+
+    afterEach(() => {
+        proxy.close();
+    });
+
+    it('Should process pages', () => {
+        session.id = 'sessionId';
+
+        session.injectable.scripts.push('/script1.js', '/script2.js');
+        session.injectable.styles.push('/styles1.css', '/styles2.css');
+
+        const options = {
+            url:     proxy.openSession('http://127.0.0.1:2000/page', session),
+            headers: {
+                accept: PAGE_ACCEPT_HEADER
+            },
+            rejectUnauthorized: false
+        };
+
+        expect(options.url).eql('https://127.0.0.1:1836/sessionId*12345/http://127.0.0.1:2000/page');
+
+        return request(options)
+            .then(body => {
+                const expected = fs.readFileSync('test/server/data/page/expected-https.html').toString();
+
+                compareCode(body, expected);
+            });
+    });
+});
